Refine types for GC threshold in SourceBufferStream
It was int, and now replaced by int64_t to be in sync with the type used
for TimeDelta.  This avoids value being truncated when the return value
of GetBufferedDurationForGarbageCollection() is large.

(cherry picked from commit 5fe17c6f697b1b0e7379a6377a9466a97818d05b)

+++ cobalt/media/filters/source_buffer_stream.cc
@@ -736,7 +736,7 @@ bool SourceBufferStream::GarbageCollectIfNeeded(DecodeTimestamp media_time,
 
   size_t bytes_to_free = 0;
 
-  int garbage_collection_duration_threshold_in_seconds =
+  int64_t garbage_collection_duration_threshold_in_seconds =
       SbMediaGetBufferGarbageCollectionDurationThreshold() / kSbTimeSecond;
   // Check if we're under or at the memory/duration limit.
   const auto kGcDurationThresholdInMilliseconds =
